requirements.txt
ultralytics
torch
torchvision
opencv-python
numpy 
matplotlib
pillow
reportlab
pandas